pyproject.toml
[project]
name = "soft-rob-embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "mujoco>=3.2.6",
]
